fix add-transaction build issue
